tests/wpt/tests/html/semantics/scripting-1/the-script-element/text-module/module.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Text modules</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
const t = async_test();
</script>
<script type="module" onerror="t.step(() => assert_unreached(event))">
import v from "./file.txt" with { type: "text" };
t.step(() => {
  assert_equals(typeof v, "string");
  assert_equals(v, "text file\n");
  t.done();
});
</script>
